First integration of HTX in H1 state
