+++ lib/Cargo.toml
@@ -57,6 +57,7 @@ time = "^0.3.20"
 url = "^2.3.1"
 webpki = "^0.22.0"
 x509-parser = "^0.15.0"
+htx = { version = "0.3.1", git = "https://github.com/CleverCloud/htx"}
 
 sozu-command-lib = { path = "../command" }
