+++ lib/src/protocol/mod.rs
@@ -51,7 +51,8 @@ macro_rules! empty (
 );
 
 pub mod h2;
-pub mod http;
+// pub mod http;
+pub mod htx_h1;
 pub mod pipe;
 pub mod proxy_protocol;
 pub mod rustls;
@@ -67,6 +68,7 @@ use crate::{
 
 pub use self::{
     http::{Http, StickySession},
+    htx_h1 as http,
     pipe::Pipe,
     proxy_protocol::send::SendProxyProtocol,
     rustls::TlsHandshake,
