+++ lib/src/tcp.rs
@@ -223,7 +223,7 @@ impl TcpSession {
         if let Some(backend_id) = self.metrics.backend_id.as_ref() {
             if let Some(backend_response_time) = self.metrics.backend_response_time() {
                 record_backend_metrics!(
-                    cluster_id,
+                    &cluster_id,
                     backend_id,
                     backend_response_time.whole_milliseconds(),
                     self.metrics.backend_response_time(),
